==> bikenwgrowth-master/code/06_video_generation.ipynb <==
# Step 6 - Making videos of the plots
## Project: Growing Urban Bicycle Networks

This notebook takes the plots from 05_plot_results and creates videos out of them. Run a kernel which has opencv-python available!

Contact: Michael Szell (michael.szell@gmail.com)  
Created: 2020-09-02  
Last modified: 2021-03-10

## Preliminaries

### Parameters

debug = False # If True, will produce plots and/or verbose output to double-check
%run -i "../parameters/parameters.py"

### Setup

%run -i path.py

import csv
cities = {}


import cv2
import os
from tqdm.notebook import tqdm
import itertools

cities = {'copenhagen': {'nominatimstring': '', 'countryid': 'dnk', 'name': 'Copenhagen'}}

### Functions

from shapely.geometry import Polygon
import igraph as ig
%run -i functions.py

attrlist = ['length','length_attr', 'bcount_attr', 'pop_den_attr', 'bcount_pop_den_attr']

## Generate Videos

### POI based results

%run -i functions.py
# Run all parameter sets
poi_source_list = ["grid"]
prune_measure_list = ["betweenness"]
parsets_used = list(itertools.product(poi_source_list, prune_measure_list))
for poi_source, prune_measure in parsets_used:
    print(poi_source, prune_measure)
    
    for placeid, placeinfo in tqdm(cities.items(), desc = "Cities"):
        try:
            imgname = '_GTall_poi_' + poi_source + "_" + prune_measures[prune_measure]
            generate_video(placeid, imgname,'mp4')
            imgname = '_GTabstract_poi_' + poi_source + "_" + prune_measures[prune_measure]
            generate_video(placeid, imgname,'mp4')
            imgname = '_GTallcover_poi_' + poi_source + "_" + prune_measures[prune_measure]
            generate_video(placeid, imgname,'mp4')
#             imgname = '_GTalloverlapbikeable_poi_' + poi_source + "_" + prune_measures[prune_measure]
#             generate_video(placeid, imgname)
#             imgname = '_GTalloverlapbiketrack_poi_' + poi_source + "_" + prune_measures[prune_measure]
#             generate_video(placeid, imgname)
        except IndexError:
            continue
        

vid

# Run all parameter sets
poi_source_list = ["grid"]
prune_measure_list = ["betweenness"]
parsets_used = list(itertools.product(poi_source_list, prune_measure_list))
for poi_source, prune_measure in parsets_used:
    print(poi_source, prune_measure)
    
    for placeid, placeinfo in tqdm(cities.items(), desc = "Cities"):
        for attr in attrlist:
            try:
                imgname = '_GTall_poi_' + poi_source + "_" + prune_measures[prune_measure]
                generate_video_custom(placeid, imgname, attr)
                imgname = '_GTabstract_poi_' + poi_source + "_" + prune_measures[prune_measure]
                generate_video_custom(placeid, imgname, attr)
                imgname = '_GTallcover_poi_' + poi_source + "_" + prune_measures[prune_measure]
                generate_video_custom(placeid, imgname, attr)
    #             imgname = '_GTalloverlapbikeable_poi_' + poi_source + "_" + prune_measures[prune_measure]
    #             generate_video(placeid, imgname)
    #             imgname = '_GTalloverlapbiketrack_poi_' + poi_source + "_" + prune_measures[prune_measure]
    #             generate_video(placeid, imgname)
            except IndexError:
                continue
